+++ utils/records.js
@@ -6,13 +6,16 @@
 
 'use strict';
 
-const immutable                     = require('immutable');
+const Immutable                     = require('immutable');
 
 
-class Message extends immutable.Record({ hasError: false, data: {} }) {
+/**
+ * @private Base record used to build more concrete message classes like {@link SuccessMsg} and {@link ErrorMsg}.
+ */
+class Message extends Immutable.Record({ hasError: false, data: {} }) {
   constructor(data) {
-    data.data = immutable.fromJS(data.data, function (key, value) {
-      var isIndexed = immutable.Iterable.isIndexed(value);
+    data.data = Immutable.fromJS(data.data, function (key, value) {
+      var isIndexed = Immutable.Iterable.isIndexed(value);
       return isIndexed ? value.toList() : value.toOrderedMap();
     });
 
@@ -20,10 +23,22 @@ class Message extends immutable.Record({ hasError: false, data: {} }) {
   }
 }
 
+
+/**
+ * Record to pass a payload for success messages around.
+ *
+ * @param {Object} data
+ */
 const SuccessMsg = (data) => new Message({
   data: data
 });
 
+
+/**
+ * Record to pass a payload for error messages around.
+ *
+ * @param {Object} err
+ */
 const ErrorMsg = (err) => new Message({
   hasError: true,
   data: {
@@ -31,7 +46,9 @@ const ErrorMsg = (err) => new Message({
   }
 });
 
+
 const pkgDefaults = {
+  authorCfg: new Immutable.Map(),
   isHost: true,
   basename: null,
   name: null,
@@ -43,15 +60,29 @@ const pkgDefaults = {
   data: {} //only for sub host dep
 };
 
-const Package = immutable.Record(pkgDefaults);
 
+const Package = Immutable.Record(pkgDefaults);
+
+
+/**
+ * Record to store host generator's package data.
+ *
+ * @param {Object} pkg
+ */
 const HostGenPkg = (pkg={}) => Package(Object.assign({}, pkgDefaults, pkg, {
   isHost: true
 }));
 
+
+/**
+ * Record to store subgen's package data.
+ *
+ * @param {Object} pkg
+ */
 const SubGenPkg = (pkg={}) => Package(Object.assign({}, pkgDefaults, pkg, {
   isHost: false
 }));
+
 
 module.exports = {
   SuccessMsg,